integracao das balas com o sistema de valores

+++ MicroJAM/Assets/Scripts/OnHitDamage.cs
@@ -6,7 +6,7 @@ public class OnHitDamage : MonoBehaviour
 {
     [Header("Configurações")]
     public string targetTag = "Enemy"; // Tag do alvo (por exemplo, "enemy")
-    public int damage = 10; // Dano causado ao inimigo
+    public float damage = 10; // Dano causado ao inimigo
 
     // Esse método é chamado quando o objeto com este script colide com outro objeto (via trigger)
     void OnTriggerEnter2D(Collider2D other)
